requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
joblib
pydantic
boto3
python-dotenv
